+++ src/j4pr.jl
@@ -141,6 +141,7 @@ module j4pr
 		RandomSubspace, randomsubspace,							# Random sub-space ensemble
 		AdaBoost, adaboost,								# AdaBoost ensemble
 		DecisionStump, stump, stumpr,							# Decision stump classifier and regressor
+		NetworkLearner,									# Network learning
 
 		# [lib/unsup]
 		kernelize!, kernelize, kernel							# Construct kernels
@@ -209,6 +210,7 @@ module j4pr
 		include("lib/sup/randomsubspace.jl")						# Random subspace ensemble framework
 		include("lib/sup/adaboost.jl")							# AdaBoost ensemble framework
 		include("lib/sup/stump.jl")							# Decision stump classifier and regressor
+		include("lib/sup/networklearner/NetworkLearner.jl")				# Network learning 
 	
 	# Some packages do not work for Julia > 0.6
 	if v"0.6" <= VERSION < v"0.7-"